+++ lib/cuke_modeler/models/background.rb
@@ -4,11 +4,16 @@ module CukeModeler
 
   class Background < Model
 
+    include Parsing
     include Parsed
     include Named
     include Described
     include Stepped
     include Sourceable
+
+
+    # The background's keyword
+    attr_accessor :keyword
 
 
     # Creates a new Background object and, if *source_text* is provided, populates
@@ -41,7 +46,7 @@ module CukeModeler
     def to_s
       text = ''
 
-      text << "Background:#{name_output_string}"
+      text << "#{@keyword}:#{name_output_string}"
       text << "\n" + description_output_string unless (description.nil? || description.empty?)
       text << "\n" unless (steps.empty? || description.nil? || description.empty?)
       text << "\n" + steps_output_string unless steps.empty?
@@ -54,7 +59,7 @@ module CukeModeler
 
 
     def parse_source(source_text)
-      base_file_string = "Feature: Fake feature to parse\n"
+      base_file_string = "#{dialect_feature_keyword}: Fake feature to parse\n"
       source_text = base_file_string + source_text
 
       parsed_file = Parsing::parse_text(source_text, 'cuke_modeler_stand_alone_background.feature')