+++ lib/cuke_modeler/models/example.rb
@@ -4,12 +4,16 @@ module CukeModeler
 
   class Example < Model
 
+    include Parsing
     include Parsed
     include Named
     include Described
     include Sourceable
     include Taggable
 
+
+    # The example's keyword
+    attr_accessor :keyword
 
     # The row models in the example table
     attr_accessor :rows
@@ -101,10 +105,10 @@ module CukeModeler
       text = ''
 
       text << tag_output_string + "\n" unless tags.empty?
-      text << "Examples:#{name_output_string}"
+      text << "#{@keyword}:#{name_output_string}"
       text << "\n" + description_output_string unless (description.nil? || description.empty?)
-      text << "\n" unless (description.nil? || description.empty?)
-      text << "\n" + parameters_output_string
+      text << "\n" unless (rows.empty? || description.nil? || description.empty?)
+      text << "\n" + parameters_output_string if parameter_row
       text << "\n" + rows_output_string unless argument_rows.empty?
 
       text
@@ -115,7 +119,7 @@ module CukeModeler
 
 
     def parse_source(source_text)
-      base_file_string = "Feature: Fake feature to parse\nScenario Outline:\n* fake step\n"
+      base_file_string = "#{dialect_feature_keyword}: Fake feature to parse\n#{dialect_outline_keyword}:\n#{dialect_step_keyword} fake step\n"
       source_text = base_file_string + source_text
 
       parsed_file = Parsing::parse_text(source_text, 'cuke_modeler_stand_alone_example.feature')