+++ lib/cuke_modeler/models/row.rb
@@ -5,6 +5,7 @@ module CukeModeler
   class Row < Model
 
     include Sourceable
+    include Parsing
     include Parsed
 
     # The cell models that make up the row
@@ -37,7 +38,7 @@ module CukeModeler
 
 
     def parse_source(source_text)
-      base_file_string = "Feature: Fake feature to parse\nScenario:\n* fake step\n"
+      base_file_string = "#{dialect_feature_keyword}: Fake feature to parse\n#{dialect_scenario_keyword}:\n#{dialect_step_keyword} fake step\n"
       source_text = base_file_string + source_text
 
       parsed_file = Parsing::parse_text(source_text, 'cuke_modeler_stand_alone_row.feature')