+++ lib/cuke_modeler/models/table.rb
@@ -4,6 +4,7 @@ module CukeModeler
 
   class Table < Model
 
+    include Parsing
     include Parsed
     include Sourceable
 
@@ -41,7 +42,7 @@ module CukeModeler
 
 
     def parse_source(source_text)
-      base_file_string = "Feature:\nScenario:\n* step\n"
+      base_file_string = "#{dialect_feature_keyword}:\n#{dialect_scenario_keyword}:\n#{dialect_step_keyword} step\n"
       source_text = base_file_string + source_text
 
       parsed_file = Parsing::parse_text(source_text, 'cuke_modeler_stand_alone_table.feature')