+++ lib/cuke_modeler/models/tag.rb
@@ -4,6 +4,7 @@ module CukeModeler
 
   class Tag < Model
 
+    include Parsing
     include Parsed
     include Sourceable
 
@@ -34,7 +35,7 @@ module CukeModeler
 
 
     def parse_source(source_text)
-      base_file_string = "\nFeature: Fake feature to parse"
+      base_file_string = "\n#{dialect_feature_keyword}: Fake feature to parse"
       source_text = source_text + base_file_string
 
       parsed_file = Parsing::parse_text(source_text, 'cuke_modeler_stand_alone_tag.feature')