+++ testing/rspec/spec/spec_helper.rb
@@ -21,6 +21,40 @@ require "#{this_dir}/unit/shared/bare_bones_models_unit_specs"
 require "#{this_dir}/unit/shared/prepopulated_models_unit_specs"
 require "#{this_dir}/unit/shared/sourced_models_unit_specs"
 require "#{this_dir}/unit/shared/parsed_models_unit_specs"
+require "#{this_dir}/unit/shared/keyworded_models_unit_specs"
+
+require "#{this_dir}/../../dialect_helper"
+
+
+# Use a fake dialect for testing in order to avoid hard coded language assumptions in the
+# implementation. Only possible with newer versions of Gherkin.
+if Gem.loaded_specs['gherkin'].version.version[/^2/]
+  CukeModeler::DialectHelper.set_dialect(Gherkin::I18n::LANGUAGES['en'])
+  CukeModeler::Parsing.dialect = 'en'
+else
+  dialect_file_path = "#{this_dir}/../../test_languages.json"
+  test_dialects = JSON.parse File.open(dialect_file_path, 'r:UTF-8').read
+
+  Gherkin::DIALECTS.merge!(test_dialects)
+
+
+  # Making the test dialect the default dialect so that language headers aren't
+  # needed for all of the test code.
+  module Gherkin
+    class Parser
+
+      alias_method :original_parse, :parse
+
+      def parse(token_scanner, token_matcher = TokenMatcher.new('cm-test'))
+        original_parse(token_scanner, token_matcher)
+      end
+
+    end
+  end
+
+  CukeModeler::DialectHelper.set_dialect(test_dialects['cm-test'])
+  CukeModeler::Parsing.dialect = 'cm-test'
+end
 
 
 RSpec.configure do |config|
@@ -40,6 +74,15 @@ RSpec.configure do |config|
   end
 
   config.before(:all) do
+    @feature_keyword = CukeModeler::DialectHelper.feature_keyword
+    @background_keyword = CukeModeler::DialectHelper.background_keyword
+    @scenario_keyword = CukeModeler::DialectHelper.scenario_keyword
+    @outline_keyword = CukeModeler::DialectHelper.outline_keyword
+    @example_keyword = CukeModeler::DialectHelper.example_keyword
+    @step_keyword = CukeModeler::DialectHelper.step_keyword
+    @given_keyword = CukeModeler::DialectHelper.given_keyword
+    @then_keyword = CukeModeler::DialectHelper.then_keyword
+
     @default_file_directory = "#{this_dir}/temp_files"
     @default_feature_file_name = 'test_feature.feature'
   end