+++ testing/rspec/spec/unit/example_unit_spec.rb
@@ -9,6 +9,7 @@ describe 'Example, Unit', :unit_test => true do
   describe 'common behavior' do
 
     it_should_behave_like 'a model'
+    it_should_behave_like 'a keyworded model'
     it_should_behave_like 'a named model'
     it_should_behave_like 'a described model'
     it_should_behave_like 'a tagged model'
@@ -108,6 +109,12 @@ describe 'Example, Unit', :unit_test => true do
 
         it 'can output an empty example' do
           expect { example.to_s }.to_not raise_error
+        end
+
+        it 'can output an example that has only a keyword' do
+          example.keyword = 'foo'
+
+          expect(example.to_s).to eq('foo:')
         end
 
         it 'can output an example that has only a name' do