+++ testing/rspec/spec/unit/feature_unit_spec.rb
@@ -10,6 +10,7 @@ describe 'Feature, Unit', :unit_test => true do
   describe 'common behavior' do
 
     it_should_behave_like 'a model'
+    it_should_behave_like 'a keyworded model'
     it_should_behave_like 'a named model'
     it_should_behave_like 'a described model'
     it_should_behave_like 'a tagged model'
@@ -132,6 +133,12 @@ describe 'Feature, Unit', :unit_test => true do
 
         it 'can output an empty feature' do
           expect { feature.to_s }.to_not raise_error
+        end
+
+        it 'can output a feature that has only a keyword' do
+          feature.keyword = 'foo'
+
+          expect(feature.to_s).to eq('foo:')
         end
 
         it 'can output a feature that has only a name' do