+++ testing/rspec/spec/unit/outline_unit_spec.rb
@@ -10,6 +10,7 @@ describe 'Outline, Unit', :unit_test => true do
   describe 'common behavior' do
 
     it_should_behave_like 'a model'
+    it_should_behave_like 'a keyworded model'
     it_should_behave_like 'a named model'
     it_should_behave_like 'a described model'
     it_should_behave_like 'a stepped model'
@@ -92,6 +93,12 @@ describe 'Outline, Unit', :unit_test => true do
 
         it 'can output an empty outline' do
           expect { outline.to_s }.to_not raise_error
+        end
+
+        it 'can output an outline that has only a keyword' do
+          outline.keyword = 'foo'
+
+          expect(outline.to_s).to eq('foo:')
         end
 
         it 'can output an outline that has only a name' do