+++ testing/rspec/spec/unit/scenario_unit_spec.rb
@@ -10,6 +10,7 @@ describe 'Scenario, Unit', :unit_test => true do
   describe 'common behavior' do
 
     it_should_behave_like 'a model'
+    it_should_behave_like 'a keyworded model'
     it_should_behave_like 'a named model'
     it_should_behave_like 'a described model'
     it_should_behave_like 'a stepped model'
@@ -61,6 +62,12 @@ describe 'Scenario, Unit', :unit_test => true do
 
         it 'can output an empty scenario' do
           expect { scenario.to_s }.to_not raise_error
+        end
+
+        it 'can output a scenario that has only a keyword' do
+          scenario.keyword = 'foo'
+
+          expect(scenario.to_s).to eq('foo:')
         end
 
         it 'can output a scenario that has only a name' do